Update README.md
Camtasia Studio Crack  Is a useful program that lets you make videos of what’s happening on your screen in any program that’s running on your computer. After the videos are made, they can be shared with anyone as tutorials. It is easy to set up and will be done quickly.